Review + Rating system
Limited Reviews to 100 ch. and ratings to 1-5
Reviews will not appear before they are approved by an employee
BookController now calculates average rating, view makes format proper
We need identity to finish the reviews system so the code is able to reference the currently logged in user.

+++ Group_18_Final_Project/Group_18_Final_Project/Models/Book.cs
@@ -13,7 +13,7 @@ namespace Group_18_Final_Project.Models
         public String Author { get; set; }
         public Int32 UniqueID { get; set; }
         public Int32 TimesPurchased { get; set; }
-        public Int32 AverageRating { get; set; }
+        public Decimal AverageRating { get; set; }
         public Int32 CopiesOnHand { get; set; }
         public Decimal BookPrice { get; set; }
         public Decimal WholesalePrice { get; set; }
